meltano/meltano_project/requirements.txt
psycopg2-binary
pandas
python-dotenv
sqlalchemy
openpyxl